Remove jackson module
