--- infrastructure/jackson/build.gradle
@@ -1,8 +1,0 @@
-dependencies {
-    api 'com.fasterxml.jackson.core:jackson-databind'
-    api 'io.swagger.core.v3:swagger-annotations'
-
-    implementation project(':infrastructure:bytes')
-
-    implementation 'io.tmio:tuweni-units'
-}
