--- infrastructure/jackson/src/main/java/tech/pegasys/teku/infrastructure/jackson/deserializers/bytes/Bytes48KeyDeserializer.java
@@ -1,33 +1,0 @@
-/*
- * Copyright Consensys Software Inc., 2022
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.infrastructure.jackson.deserializers.bytes;
-
-import com.fasterxml.jackson.core.JacksonException;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.KeyDeserializer;
-import org.apache.tuweni.bytes.Bytes48;
-
-public class Bytes48KeyDeserializer extends KeyDeserializer {
-
-  @Override
-  public Object deserializeKey(final String key, final DeserializationContext ctxt)
-      throws JacksonException {
-    try {
-      return Bytes48.fromHexStringStrict(key);
-    } catch (RuntimeException ex) {
-      throw new JsonMappingException(ctxt.getParser(), "Invalid Public Key", ex);
-    }
-  }
-}
