--- infrastructure/jackson/src/main/java/tech/pegasys/teku/infrastructure/jackson/deserializers/bytes/BytesSerializer.java
@@ -1,30 +1,0 @@
-/*
- * Copyright Consensys Software Inc., 2022
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.infrastructure.jackson.deserializers.bytes;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import java.io.IOException;
-import java.util.Locale;
-import org.apache.tuweni.bytes.Bytes;
-
-public class BytesSerializer extends JsonSerializer<Bytes> {
-  @Override
-  public void serialize(
-      final Bytes value, final JsonGenerator gen, final SerializerProvider provider)
-      throws IOException {
-    gen.writeString(value.toHexString().toLowerCase(Locale.ROOT));
-  }
-}
