+++ settings.gradle
@@ -33,7 +33,6 @@ include 'infrastructure:events'
 include 'infrastructure:exceptions'
 include 'infrastructure:http'
 include 'infrastructure:io'
-include 'infrastructure:jackson'
 include 'infrastructure:json'
 include 'infrastructure:kzg'
 include 'infrastructure:metrics'
