+++ validator/client/build.gradle
@@ -4,7 +4,6 @@ dependencies {
   implementation project(':infrastructure:crypto')
   implementation project(':infrastructure:exceptions')
   implementation project(':infrastructure:io')
-  implementation project(':infrastructure:jackson')
   implementation project(':infrastructure:metrics')
   implementation project(':infrastructure:http')
   implementation project(':ethereum:json-types')
